Features: Add some synthetic features and fix some bugs.
The linespaces and font size data was broken, fixed here.

Added new synthetic features "scaled margins". The top and bottom
"scaled margin" is the corresponding margin divided by the
line_height (or if no line_height, because it's a single line, 120%
of the font size). The left and right "scaled margin" is the
corresponding margin divided by the char size.