Add new 'dodgy_paragraph_breaks' feature.
We detect the space at the end of each line, and compare it
to the size of the first word on the next line.

If that first word (plus a space) would have fitted into the
gap at the end of the previous line, and the end of the previous
line ended with a non-alphanumeric character, then it looks
like that might have been the end of a paragraph.

The fact the block doesn't stop there suggests that we've
failed to break for a paragraph - hence 'dodgy_paragraph_breaks'.

The logic for this is lifted from MuPDFs paragraph breaking
code, at the suggestion of Youchang Kim.