Refactor, and add contextual above/below features.
Refactor the code out of being one monolithic procedure, and add
some new features, intended to give context above and below the
selected block.

We look at the line immediately above the selected region, and
score it with how much it might be part of the same block.

If it's a title, and we aren't (for some definition of title),
we score 0.

If it's a different sized font, we score 0.

If it's looks plausibly like it was indented at the end, then
it's probably a different paragraph, so we score 0.

Otherwise we score how far within our margin it was (so very far
away scores low, and close scores almost 50).

Similarly for the line below.