Add 4 more features, for left/right alignment with baselines.
Look for blocks that align with the 'baseline' of the current
block.

The baseline of the current block is defined to be the baseline
of the first char of the highest horizontally-moving line with
the highest fontsize.

This means that if a region covers a paragraph, it'll be the
baseline of the topmost line used. If the line includes
subscripts/superscripts, these are typically lower fontsizes
so they should not be picked.