More rafting changes.
Most of the time, when we raft a new plank, it'll attach to the
last raft we have. Test that one first.

Also, when we spot that a plank bridges 2 rafts, the resulting
merged raft (represented, as it is, simply by a rectangle) can
then bridge multiple other rafts. Thus we need to iterate several
times.

For example:

         +-----+
   +---+ | 0   |
   | 1 | +-----+
   +---+------+
       | 2    |
       +------+

Rectangle 2 is spotted as abutting rectangle 1. The combined
rectangle from them, then overlaps rectangle 0 too.

Some useful debug code was used to help fix this problem; this
is now left in, but disabled by #ifdefs.